Cleanup common code
